add refactor category
