--- src/data/categories.ts
@@ -1,19 +1,0 @@
-export const categories = [
-  { key: "all" as const, label: "Todos" },
-  { key: "medicamentos" as const, label: "Medicamentos" },
-  { key: "vitaminas" as const, label: "Vitaminas" },
-  { key: "cuidado-personal" as const, label: "Cuidado Personal" },
-  { key: "suplementos" as const, label: "Suplementos" },
-  { key: "belleza" as const, label: "Belleza" },
-  { key: "ropa" as const, label: "Ropa" },
-  { key: "zapatos" as const, label: "Zapatos" },
-  { key: "telefonos" as const, label: "Teléfonos" },
-  { key: "computadoras" as const, label: "Computadoras" },
-];
-
-export const sortOptions = [
-  { key: "precio", label: "Precio" },
-  { key: "relevancia", label: "Relevancia" },
-  { key: "popularidad", label: "Popularidad" },
-  { key: "descuento", label: "Descuento" },
-];
